ubiformat/ubinize: fix more 64-bit image sequence number confusion
The sequence number is only 32 bit as the strtoul() function.

+++ ubi-utils/src/ubiformat.c
@@ -134,7 +134,7 @@ static int parse_opt(int argc, char * const argv[])
 	while (1) {
 		int key;
 		char *endp;
-		unsigned long long image_seq;
+		unsigned long int image_seq;
 
 		key = getopt_long(argc, argv, "nh?Vyqve:x:s:O:f:S:", long_options, NULL);
 		if (key == -1)
